Move ResultT.h from arangod/Cluster to lib/Basics
This is the correct home for this code.
