+++ arangod/Cluster/SynchronizeShard.h
@@ -27,7 +27,7 @@
 
 #include "ActionBase.h"
 #include "ActionDescription.h"
-#include "Cluster/ResultT.h"
+#include "Basics/ResultT.h"
 #include "VocBase/voc-types.h"
 
 #include <chrono>
