+++ arangod/RestHandler/RestSupervisionStateHandler.cpp
@@ -26,7 +26,7 @@
 
 #include "Agency/AsyncAgencyComm.h"
 #include "Cluster/AgencyPaths.h"
-#include "Cluster/ResultT.h"
+#include "Basics/ResultT.h"
 #include "GeneralServer/GeneralServer.h"
 #include "GeneralServer/GeneralServerFeature.h"
 #include "GeneralServer/RestHandlerFactory.h"
