+++ arangod/RestHandler/RestTestHandler.cpp
@@ -22,7 +22,7 @@
 
 #include "RestTestHandler.h"
 
-#include "Cluster/ResultT.h"
+#include "Basics/ResultT.h"
 #include "GeneralServer/GeneralServer.h"
 #include "GeneralServer/GeneralServerFeature.h"
 #include "GeneralServer/RestHandlerFactory.h"
